moved all js files in to corresponding folders
