+++ front-end/src/pages/MainPages/About.js
