+++ front-end/src/pages/OnAllPages/Footer.js
