+++ front-end/src/pages/OnAllPages/Navbar.js
@@ -1,29 +1,29 @@
 import React from "react";
 import { useEffect, useState } from "react";
 import { BrowserRouter, Link, Route, Routes } from "react-router-dom";
-import logo from "../logo.png";
+import logo from "../../logo.png";
 
-import Home from "./Home";
-import About from "./About";
-import Contact from "./Contact";
-import Services from "./Services";
-import Team from "./Team";
-import Login from "./Login";
-import Signin from "./Signin";
-import Checkout from "./Checkout";
-import Products from "./Products";
-import ClientPics from "./ClientPics";
-import Admin from "./Admin";
-import Cart from "./Cart";
-import ClientStories from "./ClientStories";
-import Confirm from "./Confirm";
-import EditRoles from "./EditRoles";
-import NotFound from "./NotFound";
-import EditImages from "./EditImages";
-import ShowClientPics from "./ShowClientPics";
-import BookingForm from "./BookingForm";
-import BookingInfo from "./BookingInfo";
-import ReviewOrder from "./ReviewOrder";
+import Home from "../MainPages/Home";
+import About from "../MainPages/About";
+import Contact from "../MainPages/Contact";
+import Services from "../MainPages/Services";
+import Team from "../MainPages/Team";
+import Login from "../MainPages/Login";
+import Signin from "../MainPages/Signin";
+import Checkout from "../CheckoutProcess/Checkout";
+import Products from "../CheckoutProcess/Products";
+import ClientPics from "../Client/ClientPics";
+import Admin from "../Admin/Admin";
+import Cart from "../CheckoutProcess/Cart";
+import ClientStories from "../MainPages/ClientStories";
+import Confirm from "../CheckoutProcess/Confirm";
+import EditRoles from "../Admin/EditRoles";
+import NotFound from "../ErrorPages/NotFound";
+import EditImages from "../Admin/EditImages";
+import ShowClientPics from "../Admin/ShowClientPics";
+import BookingForm from "../MainPages/BookingForm";
+import BookingInfo from "../Admin/BookingInfo";
+import ReviewOrder from "../CheckoutProcess/ReviewOrder";
 
 const Navbar = () => {
 	const [loggedIn, setLoggedIn] = useState(false);
